Create Question game.js
This game will pick a random question and it will present the question to the user. Upon receiving an answer from the user, the program will update with a message if the answer is correct or incorrect and will update the score accordingly. The process will continue until the user enters 'EXIT'.